test(core/message): add test files
