+++ packages/core/src/tests/modal/modal.e2e.ts
@@ -9,13 +9,7 @@
 
 import { expect } from '@playwright/test';
 import { regressionTest, test } from '@utils/test';
-import { showMessage } from 'src/components/utils/modal/message';
 import { IxModalSize } from 'src/components';
-declare global {
-  interface Window {
-    showMessage: typeof showMessage;
-  }
-}
 
 regressionTest.describe('modal', () => {
   regressionTest('basic', async ({ page }) => {
@@ -135,45 +129,6 @@ screenWidths.forEach((size) => {
   });
 });
 
-screenWidths.forEach((size) => {
-  test(`message size ${size}`, async ({ page, mount }) => {
-    await mount(``);
-
-    await page.evaluate(() => {
-      return new Promise<void>((resolve) => {
-        const script = document.createElement('script');
-        script.type = 'module';
-        script.innerHTML = `
-        import * as ix from 'http://127.0.0.1:8080/www/build/index.esm.js';
-        window.showMessage = ix.showMessage;
-      `;
-
-        document.body.appendChild(script);
-        resolve();
-      });
-    });
-
-    await page.waitForTimeout(250);
-
-    await page.evaluate(
-      ({ size }) => {
-        window.showMessage({
-          messageTitle: 'Example title',
-          message: 'message',
-          icon: 'info',
-          size: size,
-          centered: true,
-          actions: [],
-        });
-      },
-      { size }
-    );
-
-    await page.waitForTimeout(250);
-
-    expect(await page.screenshot({ fullPage: true })).toMatchSnapshot();
-  });
-});
 
 test('modal should show centered', async ({ mount, page }) => {
   await mount(`
